Created SearchTextfield Design
SearchTextField design is created with backgroundColor "#F1F6F9" and placeholder "Find your best Salon"
